Added User Profile wireframe
Added two wireframes for the user profile. One logged in as the user and
one not logged in to show the difference. The 'gamification' aspects
whereby a user may earn a certain title (for instance 'Captain Common')
is based on actions performed by this user. It is definitely
something which needs to be discussed. The general idea is to have some
sort of feedback mechanism to 'seduce' users into putting (more &
recurrent) energy into the Catalog using ego as a motivational driver.