Added metadata editor wireframe proposal
The metadata editor wireframe which I have created does not take a
specific media type into consideration in the wireframe. Instead it
shows all metadata regardless of if this is part of the media type. In
practice the metadata editor should be contextually aware of the media
type and only show those fields relevant for the media type. E.g. a
framerate is useless when the work itself is an image or text document.
I've chosen to follow the W3 Media ontology almost to the letter.
Deviating only there where other terms made more sense to to the general
public. Also I've altered the order for the rights metadata since I'd
consider this more important than relations in the case of most
metadata. I suggest that the final interface will allow the user to use
'folds' and only display the metadata properties he or she is interested
in.

Note:
At this moment the metadata editor is missing Catalog specific
metadata, such as the uploader, time of addition to the Catalog and
such. I think we need to discuss if this should be part of the metadata
one can edit with the editor or should be considered system metadata and
therefor only relevant for the Catalog system.